v0.1.1 Updated interface to return single function
Module interface no longer returns a prototypal object but instead
returns a single function that generates a function.

There didn't seem to be a need to force users of the module to
instantiate a new object, run the setup functions and then use the
object to then call the generate function. Instead, you run the
function, passing the 2 parameters necessary for the object setup and
then it returns the balancingUrl.generateUrl function.

Additionally, the pre-built route types are available on the object if
needed (eg. balancingUrl.routeTypes.RANDOM,
balancingUrl.routeTypes.ROUND_ROBIN) as part of the setup.